+++ components/medication/form/ReminderForm.tsx
@@ -1,4 +1,4 @@
-import { Formik, type FormikErrors } from "formik";
+import { Formik } from "formik";
 import { useState } from "react";
 import {
   Keyboard,
@@ -11,14 +11,12 @@ import { Button } from "react-native-paper";
 import { CronPicker } from "./CronPicker";
 import { FormInput } from "./FormInput";
 import { MedicationPicker } from "./MedicationPicker";
+import { deserializeCreate, deserializeEdit, validate } from "./data";
 import type { ReminderFormValues } from "./types";
 
-import { useI18n, type Translator } from "@/common/i18n";
+import { useI18n } from "@/common/i18n";
 import { sty } from "@/common/styles";
-import {
-  type ReminderCreateData,
-  type ReminderEditData,
-} from "@/logic/medication";
+import type { ReminderCreateData, ReminderEditData } from "@/logic/medication";
 
 type ReminderFormProps = {
   kind: "create" | "edit";
@@ -114,59 +112,3 @@ const styles = sty.create({
     marginTop: 32,
   },
 });
-
-const parseNum = (value: string): number | null => {
-  const num = parseInt(value, 10);
-  return isNaN(num) ? null : num;
-};
-
-const deserializeCreate = (values: ReminderFormValues): ReminderCreateData => ({
-  ...deserializeEdit(values),
-  medicationName: values.medicationName,
-  medicationAmountInPackage: parseNum(values.medicationAmountInPackage),
-  amountUnit: values.amountUnit.trim().length > 0 ? values.amountUnit : null,
-  cron: values.cron.trim().length > 0 ? values.cron : null,
-  description: values.description.trim().length > 0 ? values.description : null,
-});
-
-const deserializeEdit = (values: ReminderFormValues): ReminderEditData => ({
-  amountPerIntake: parseNum(values.amountPerIntake) ?? 0,
-  amountOwned: parseNum(values.amountOwned),
-  cron: values.cron.trim().length > 0 ? values.cron : null,
-  description: values.description.trim().length > 0 ? values.description : null,
-});
-
-const validate = (
-  values: ReminderFormValues,
-  t: Translator,
-): FormikErrors<ReminderFormValues> => {
-  const errors: FormikErrors<ReminderFormValues> = {};
-  if (values.medicationName.length === 0) {
-    errors.medicationName = t("medication.form.medicationNameRequired");
-  }
-
-  const medicationAmountInPackage = parseNum(values.medicationAmountInPackage);
-  if (
-    values.medicationAmountInPackage.trim().length > 0 &&
-    (!medicationAmountInPackage || medicationAmountInPackage < 0)
-  ) {
-    errors.medicationAmountInPackage = t(
-      "medication.form.medicationAmountInPackageNonnegative",
-    );
-  }
-
-  const amountPerIntake = parseNum(values.amountPerIntake);
-  if (!amountPerIntake || amountPerIntake <= 0) {
-    errors.amountPerIntake = t("medication.form.amountPerIntakeRequired");
-  }
-
-  const amountOwned = parseNum(values.amountOwned);
-  if (
-    values.amountOwned.trim().length > 0 &&
-    (!amountOwned || amountOwned < 0)
-  ) {
-    errors.amountOwned = t("medication.form.amountOwnedNonnegative");
-  }
-
-  return errors;
-};
